added target in makefile for building Windows artifacts and updated github workflow to run the same
In order to build Windows artifacts, we have added a target for the same in Makefile. We are using a cross-compiler to build Windows artifacts.

Also, we have added this target in the Github action.

To ensure that customers building their own plugins can build Windows binaries for different architectures and using different cross-compilers, these build settings are customizable.